+++ src/components/login/LoggedInMenu.jsx
@@ -9,8 +9,8 @@ const LoggedInMenu = ({ onLogout }) => (
       </Link>
     </li>
     <li>
-      <Link to="/dashboard" className="page-scroll">
-        Dashboard
+      <Link to="/business" className="page-scroll">
+        Business
       </Link>
     </li>
     <li>
@@ -22,9 +22,9 @@ const LoggedInMenu = ({ onLogout }) => (
         <i className="glyphicon glyphicon-user"></i>
       </Link>
       <ul className="dropdown-menu">
-        <li>
+        {/* <li>
           <Link to="dashboard">Dashboard</Link>
-        </li>
+        </li> */}
         <li>
           <Link to="profile">Profile</Link>
         </li>
