+++ src/components/login/VisitorMenu.jsx
@@ -6,7 +6,7 @@ export default props => <ul className="nav navbar-nav navbar-center dropdown ">
     <Link to="/home" className="page-scroll">Home</Link>
 </li>
 <li>
-    <Link to="/register" className="page-scroll">Register</Link>
+    <Link to="/register" className="page-scroll">Sign Up</Link>
 </li>
 <li>
     <Link to="/login" className="page-scroll">Login</Link>
